pois: lazy building of POIs
